Removed Accidental Overlap, Tightened Deadband
Still suffers from shoot through. I believe the Always On/Off PWM code
is the culprit and is allowing shoot through.
-Waveforms that transistion lower are given first chance at being set
low, this avoids accidental overlap if the waveform transitions and
locks in a double buffer setting for a cycle before self correcting
-Deadband can be semi-safely reduced down to 1 cycle, I chose 2 cycles
to be 100% sure of no overlap
-I redid the table so that it took the new deadbands into account. as
well as a minimum on time of 2 cycles (ensures it can get fully on)
-Ensured the advancing code was consistent and allowed full scale
adjustments
-The size of the array is now specified with the variable, so if an
array of incorrect size is inserted, hopefully it will through a compile
warning/error
-Cleaned up old comments that are no longer relevant