A2D readings fixed
Skipping an item in the scan skips the A2D Buffer (unbeknownst to me). I
was reading and recording correctly from AN0 to AN3, but after that
everything was shifted by two registers.
-A2D code is reading the shifted registers
-Requires an interupt on the 16th sample/convert sequence instead of
14th due to the aforementioned issue
-The current sense amplifier pins were lifted pending a redesign, as
such the reading will only return -1 until it is fixed
-Switched Grounds are now an always low open-drain output
-HOA/LOA/HOB/LOB HiV/HiA are now back in there proper locations (Port
ordered) in the pin initialization
-LED indicators are now used to show LOA on the HiV and HiA inverters so
that I can watch the frequency shifting comparatively
-A2D scan of temperature is now done off a single sample
-High Current bridge is hard coded to 60Hz, the output inverter ramps
from 20Hz to 60Hz (more or less... The timings are off so I have
compensated)
-Commented out PGC & PGD pins from the pin initialization as it is not
required and misleads anyone reading the code into thinking they are
used
-Cleaned up various comments