Added SLR to Hi-Current measurement
The measurement is very noisy and has many spikes due to circuit
ringing. The largest deviation I measured was +15A from an average of
2.4A.
Box car averaging slowed the result down to unacceptable speeds by the
time the averaging was having a noticeable effect.
A rolling average was not considered.
Slew Rate Limiting was ultimately chosen. It was quick to implement, low
overhead, and gave great results.