Added SRL to LoV sensing
Vsolar was mostly stable, but benefitted from a small amount of SLR.
The negative terminal of the transformers primary side was noisy between
the average voltage and 0V, so it benefitted.
The positive terminal of the transformers primary side had voltage
spikes between average and 0V, and nearly +40V above, so it benefitted
drastically from it.
-Each channel needed to have its own formatting function that carried
with it a memory, otherwise all three interacted with each other
-Created a common function to take care of common formating tasks for
the Low Voltage sensing once it entered the unique A2D formating
function
-Increased the starting previousMeasurement for the HiI function to the
minimum readable reading which should accelerate the Slew Rate catchup
by 270º