[Untested] Fixed ADC consistently out of order
It was interrupting after 16 samples of 14 channels, which meant it
would sample all 14 channels, then the first two channels again to fill
the memory space.
The next time it would start sampling with the "next" channel which
likely is not the 1st channel.
-Added compiler definition to allow applying my fix to be optional. I
opted for this as I don't have hardware to test with right now so I
don't know if this actually fixed the issue