UNTESTED General Bugfix
There is an address error that is only present on 15Hz, so I included a
kludge to bypass that frequency so that we can continue to test with the
pot.
Included a sentinel to ensure the Frequency can never go to 0Hz (will
cause a math error)
Fixed a bug with the target output voltage being 10 times bigger than it
should have been (maybe fixing output voltage waveforms?)
Added a PWM upper limit that is 3x deadband to always allow for a
circulating current
Added a PWM lower limit that is 1x deadband to always allow for a
conducting current
onTime is checked for validity, not circ/cond current periods. The code
to make one limit affect the other was going to be brutal, so I ensured
a sanitized input instead
The code was deoptimized but made more robust by always setting all 8 of
the registers (instead of just the 2 that needed updating)