Adding two variations on an output buck stage
I created two variations of a buck output stage that will feed a
h-bridge.

In one, the voltage input remains a constant 400V. Great for startup and
acceptable for run time. However it requires a large expensive inductor.
The main benefit is a simplicity of design.

The second has a more complicated operation, but allows the use of
smaller/cheaper components at the expense of increasing system
complexity (code). The voltage input starts at 400V to deliver the
wallop of power needed to start the motor, but then reduces to 200V for
normal operation. The reduction in voltage during run allows the
inductor to be much smaller, improves efficiency as the input voltage is
close to the output. It also improves our resolution as the maximum duty
cycle is higher so we can have more steps in between min and max.

Both designs are expecting a continous waveform. It is also expecting a
large capacitor on the input to store the required starting power.